+++ SSZipArchive/SSZipArchive.m
@@ -128,7 +128,7 @@
         [delegate zipArchiveWillUnzipArchiveAtPath:path zipInfo:globalInfo];
     }
     if ([delegate respondsToSelector:@selector(zipArchiveProgressEvent:total:)]) {
-        [delegate zipArchiveProgressEvent:(NSInteger)currentPosition total:(NSInteger)fileSize];
+        [delegate zipArchiveProgressEvent:currentPosition total:fileSize];
     }
     
     NSInteger currentFileNumber = 0;
